+++ Version 1.x/Source/dooble.cc
@@ -2665,16 +2665,24 @@ void dooble::slotQuitAndRestart(void)
 		  QDir::separator() +
 		  QCoreApplication::applicationName());
 
-  rc = ::ShellExecuteA(0, "open", program.toUtf8().constData(),
-		       0, 0, SW_SHOWNORMAL);
+  rc = ::ShellExecuteA(0,
+		       "open",
+		       program.toUtf8().constData(),
+		       0,
+		       0,
+		       SW_SHOWNORMAL);
 
-  if(rc == SE_ERR_ACCESSDENIED)
+  if((int) rc == SE_ERR_ACCESSDENIED)
     /*
     ** Elevated?
     */
 
-    ::ShellExecuteA(0, "runas", program.toUtf8().constData(),
-		    0, 0, SW_SHOWNORMAL);
+    ::ShellExecuteA(0,
+		    "runas",
+		    program.toUtf8().constData(),
+		    0,
+		    0,
+		    SW_SHOWNORMAL);
 #else
   QProcess::startDetached(QCoreApplication::applicationDirPath() +
 			  QDir::separator() +
